// frontend/entrega2/src/models/entities/professional-model.ts
import { OccupationsModel } from 'models/entities/occupation-model';
import { AreaModel } from 'models/entities/area-model';
import { CostCentreModel } from 'models/entities/cost-centre-model';
import { MasterValueModel } from './master-value-model';
import { CityModel } from './city-model';
import { DepartmentModel } from './department-model';

export class ProfessionalModel {
    id: number;
    documentNumber: string;
    idType: MasterValueModel; 
    photoUrl: string;
    code: number;
    gender: MasterValueModel;
    birthDate: string;
    name: string;
    username: string;
    password: string;
    confirmPassword: string;
    departmentExpeditionId: DepartmentModel;
    cityExpeditionId: CityModel;
    expeditionDate: Date;
    departmentBirthId: DepartmentModel;
    cityBirthId: CityModel;
    personTypeId: MasterValueModel;
    maritalStatusId: MasterValueModel;
    address: string;
    bloodType: MasterValueModel;
    departmentResidenceId: DepartmentModel;
    cityResidenceId: CityModel;
    neighborhood: string;
    phone: string;
    landline: string;
    emergencyPhone: string;
    contractingDate: Date;
    contractingTypeId: MasterValueModel;
    resolution: string;
    inscriptionTypeId: MasterValueModel;
    inscriptionType: string;
    card: string;
    occupationId: OccupationsModel;
    hasCvlac: boolean;
    cvlac: string;
    costCentreId: CostCentreModel;
    areaId: AreaModel;
    salary: number;
    benefits: boolean;
    carvedBlouse: MasterValueModel;
    trouser: MasterValueModel;
    shoeStand: MasterValueModel;
    email: string;
    stateId: MasterValueModel;

    constructor(){
        this.id = null;
        this.idType = new MasterValueModel();
        this.photoUrl = '';
        this.code = null;
        this.name = '';
        this.username = '';
        this.password = '';
        this.confirmPassword = '';
        this.gender = new MasterValueModel();
        this.departmentExpeditionId = new DepartmentModel();
        this.cityExpeditionId = new CityModel();
        this.expeditionDate = new Date();
        this.birthDate = '';
        this.landline = '';
        this.bloodType = new MasterValueModel();
        this.departmentBirthId = new DepartmentModel();
        this.cityBirthId = new CityModel();
        this.address = '';
        this.personTypeId = new MasterValueModel();
        this.maritalStatusId = new MasterValueModel();
        this.departmentResidenceId = new DepartmentModel();
        this.cityResidenceId = new CityModel();
        this.neighborhood = '';
        this.emergencyPhone = '';
        this.phone = '';
        this.contractingDate = new Date();
        this.contractingTypeId = new MasterValueModel();
        this.resolution = '';
        this.inscriptionTypeId = new MasterValueModel();
        this.inscriptionType = '';
        this.card = '';
        this.occupationId = new OccupationsModel();
        this.hasCvlac = false;
        this.cvlac = '';
        this.costCentreId = new CostCentreModel();
        this.areaId = new AreaModel();
        this.salary = 0;
        this.benefits = false;
        this.carvedBlouse = new MasterValueModel();
        this.trouser = new MasterValueModel();
        this.shoeStand = new MasterValueModel();
        this.email = '';
        this.stateId = new MasterValueModel();
    }
}